backend/requirements.txt
flask>=3.0
flask-cors
authlib
requests
python-dotenv
pyjwt
